+++ src/TextShow/Foreign/Ptr.hs
@@ -23,7 +23,7 @@ module TextShow.Foreign.Ptr (
     ) where
 
 import Data.Monoid.Compat ((<>))
-import Data.Semigroup (timesN)
+import Data.Semigroup (mtimesDefault)
 import Data.Text.Lazy.Builder (Builder, singleton)
 
 import Foreign.ForeignPtr (ForeignPtr)
@@ -87,7 +87,7 @@ instance TextShow1 Ptr where
         padOut :: Builder -> Builder
         padOut ls =
              singleton '0' <> singleton 'x'
-          <> timesN (fromIntegral . max 0 $ 2*SIZEOF_HSPTR - lengthB ls) (singleton '0')
+          <> mtimesDefault (max 0 $ 2*SIZEOF_HSPTR - lengthB ls) (singleton '0')
           <> ls
 
 instance TextShow (FunPtr a) where